getting rid of even more
